requirements.txt
aioredis==2.0.1
annotated-types==0.6.0
anyio==4.2.0
asgi-lifespan==2.1.0
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
clamd==1.0.2
click==8.1.7
fastapi==0.109.2
fastapi-limiter==0.1.6
h11==0.14.0
idna==3.6
ply==3.11
plyara==2.1.1
progress==1.6
psutil==5.9.8
pydantic==2.6.1
pydantic_core==2.16.2
PySocks==1.7.1
python-decouple==3.8
python-multipart==0.0.9
redis==5.1.0b3
requests==2.31.0
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
urllib3==2.2.1
uvicorn==0.27.1
virustotal-python==1.0.2
yara-python==4.0.5
yara-scanner==2.0.3
